set stellar characteristics

+++ GURPS Star System Generator/Star.h
@@ -13,14 +13,14 @@ private:
 	// VECTOR OF COMPANION STARS
 	bool companion{ false };			// does this star orbit a primary?
 	Star* primary{ nullptr };			// pointer to primary, if present
-	float stellar_mass{ 0 };			// expressed in solar masses
+	double stellar_mass{ 0 };			// expressed in solar masses
 	std::string spectral_type{ "" };	// spectral type of star in main sequence e.g. G2
 	int temp{ 0 };						// surface temperature in kelvins
-	float l_max{ 0 };					// maximum luminosity in main sequence
+	float l_max{ 0 };					// maximum luminosity in main sequence ** -1 if not applicable **
 	float l_min{ 0 };					// minimum luminosity in main sequence
-	float m_span{ 0 };					// duration of main sequence in billions of years
-	float s_span{ 0 };					// duration of sub-giant phase in billions of years
-	float g_span{ 0 };					// duration of giant phase in billions of years
+	float m_span{ 0 };					// duration of main sequence in billions of years ** -1 if not applicable **
+	float s_span{ 0 };					// duration of sub-giant phase in billions of years ** -1 if not applicable **
+	float g_span{ 0 };					// duration of giant phase in billions of years ** -1 if not applicable **
 
 public:
 	//**************
@@ -41,7 +41,7 @@ public:
 	float get_M_Span();				// output phase lifespans
 	float get_S_Span();				
 	float get_G_Span();
-	float get_Mass();				// output value of stellar_mass
+	double get_Mass();				// output value of stellar_mass
 	Star* get_Primary();			// output pointer to primary star or nullptr
 	std::string get_Spectral_Type();// output spectral type
 	int get_Temp();					// output temp in kelvins
